bugs fixed after review
