+++ src/components/app/App.tsx
@@ -1,0 +1,42 @@
+import { CSSProperties, useState } from 'react';
+
+import {
+	ArticleStateType,
+	defaultArticleState,
+} from 'src/constants/articleProps';
+import { ArticleParamsForm } from '../article-params-form';
+import { Article } from '../article';
+
+import styles from './App.module.scss';
+
+export const App = () => {
+	const [pageState, setPageState] = useState(defaultArticleState);
+
+	const handleFormSubmit = (data: ArticleStateType) => {
+		setPageState(data);
+	};
+
+	const handleFormReset = (data: ArticleStateType) => {
+		setPageState(data);
+	};
+
+	return (
+		<main
+			className={styles.main}
+			style={
+				{
+					'--font-family': pageState.fontFamilyOption.value,
+					'--font-size': pageState.fontSizeOption.value,
+					'--font-color': pageState.fontColor.value,
+					'--container-width': pageState.contentWidth.value,
+					'--bg-color': pageState.backgroundColor.value,
+				} as CSSProperties
+			}>
+			<ArticleParamsForm
+				onSubmit={handleFormSubmit}
+				onReset={handleFormReset}
+			/>
+			<Article />
+		</main>
+	);
+};
