+++ src/index.tsx
@@ -1,51 +1,12 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
+import { StrictMode } from 'react';
 
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
-import {
-	ArticleStateType,
-	defaultArticleState,
-} from './constants/articleProps';
+import { App } from './components/app';
 
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const [pageState, setPageState] = useState(defaultArticleState);
-
-	const handleFormSubmit = (data: ArticleStateType) => {
-		setPageState(data);
-	};
-
-	const handleFormReset = (data: ArticleStateType) => {
-		setPageState(data);
-	};
-
-	return (
-		<div
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': pageState.fontFamilyOption.value,
-					'--font-size': pageState.fontSizeOption.value,
-					'--font-color': pageState.fontColor.value,
-					'--container-width': pageState.contentWidth.value,
-					'--bg-color': pageState.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm
-				onSubmit={handleFormSubmit}
-				onReset={handleFormReset}
-			/>
-			<Article />
-		</div>
-	);
-};
 
 root.render(
 	<StrictMode>
